# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  models.SysDepartmentAddRequest:
    properties:
      describe:
        type: string
      email:
        type: string
      leader:
        type: string
      name:
        type: string
      parentId:
        type: integer
      phone:
        type: string
      sort:
        type: integer
      status:
        type: integer
    required:
    - name
    type: object
  models.SysDepartmentDeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  models.SysDepartmentUpdateRequest:
    properties:
      describe:
        type: string
      email:
        type: string
      id:
        type: integer
      leader:
        type: string
      name:
        type: string
      parentId:
        type: integer
      phone:
        type: string
      sort:
        type: integer
      status:
        type: integer
    required:
    - id
    - name
    type: object
  models.SysDictAddRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      status:
        type: integer
    required:
    - code
    - name
    type: object
  models.SysDictDeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  models.SysDictUpdateRequest:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
    required:
    - code
    - id
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
    url: https://github.com/your-repo/gin-fast/issues
  description: 基于Gin框架的快速开发脚手架API文档
  license:
    name: MIT
    url: https://github.com/your-repo/gin-fast/blob/master/LICENSE
  termsOfService: https://github.com/your-repo/gin-fast
  title: Gin-Fast API
  version: "1.0"
paths:
...
  /sysDepartment/{id}:
    get:
      consumes:
      - application/json
      description: 根据部门ID获取部门详细信息
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回部门信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 部门ID格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 根据ID获取部门信息
      tags:
      - 部门管理
  /sysDepartment/add:
    post:
      consumes:
      - application/json
      description: 创建新部门
      parameters:
      - description: 部门信息
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/models.SysDepartmentAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 部门创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 新增部门
      tags:
      - 部门管理
  /sysDepartment/delete:
    delete:
      consumes:
      - application/json
      description: 删除部门信息
      parameters:
      - description: 部门删除请求参数
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/models.SysDepartmentDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 部门删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除部门
      tags:
      - 部门管理
  /sysDepartment/edit:
    put:
      consumes:
      - application/json
      description: 更新部门信息
      parameters:
      - description: 部门信息
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/models.SysDepartmentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 部门更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新部门
      tags:
      - 部门管理
  /sysDepartment/getDivision:
    get:
      consumes:
      - application/json
      description: 获取所有部门列表（树形结构）
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回部门列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取部门列表
      tags:
      - 部门管理
  /sysDict/{id}:
    get:
      consumes:
      - application/json
      description: 根据字典ID获取字典详细信息
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回字典信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 字典ID格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 根据ID获取字典信息
      tags:
      - 字典管理
  /sysDict/add:
    post:
      consumes:
      - application/json
      description: 创建新字典
      parameters:
      - description: 字典信息
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/models.SysDictAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 字典创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 新增字典
      tags:
      - 字典管理
  /sysDict/delete:
    delete:
      consumes:
      - application/json
      description: 删除字典
      parameters:
      - description: 字典信息
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/models.SysDictDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 字典删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除字典
      tags:
      - 字典管理
  /sysDict/edit:
    put:
      consumes:
      - application/json
      description: 更新字典信息
      parameters:
      - description: 字典信息
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/models.SysDictUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 字典更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新字典
      tags:
      - 字典管理
  /sysDict/getAllDicts:
    get:
      consumes:
      - application/json
      description: 获取所有字典数据（包含关联的字典项）
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回字典列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有字典数据
      tags:
      - 字典管理
  /sysDict/getByCode/{code}:
    get:
      consumes:
      - application/json
      description: 根据字典编码获取字典及其字典项
      parameters:
      - description: 字典编码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回字典信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 字典编码不能为空
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 根据字典编码获取字典
      tags:
      - 字典管理
  /sysDict/list:
    get:
      consumes:
      - application/json
      description: 获取字典列表，支持分页和过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNum
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 字典名称
        in: query
        name: name
        type: string
      - description: 字典编码
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回字典列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 字典列表
      tags:
      - 字典管理
...
swagger: "2.0"